movies_api/requirements.txt
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-jaeger==1.20.0
tenacity==8.2.2
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
elastic-transport==8.4.0
elasticsearch==8.9.0
fastapi==0.100.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
orjson==3.9.2
pydantic==2.1.1
pydantic-settings==2.0.1
pydantic_core==2.4.0
redis==4.6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
yarl==1.9.2
PyJWT==2.8.0